redone virtual memory structure

+++ RVinterpreter/parser/include/elf_rv32_parser.hpp
@@ -3,4 +3,4 @@
 #include <elfio/elfio.hpp>
 #include <iomanip>
 
-std::pair<int64_t, std::vector<char>> get_bin_code (std::string file_name);
+std::pair<uint32_t, std::vector<char>> get_bin_code (std::string file_name);
